Rename `HttpServer` `HttpProtocol`
