+++ include/HttpProtocol.hpp
